Remove support for "outerfaces" (#164)
Joking in the title. There was no outerface feature, but the scanner would
scan `outerface` as `interface`! LOL!

JavaScript truncates bitwise integer math to 32 bits and that caused our
actual limit in the keyword scanning algorithm that used 5 bits per letter
to be 6 characters instead of the 10 we thought we had in the "safe" integer
range. The fix is to use equivalent, but non-bitwise arithmetic.

Perf testing on my machine shows this is still faster than not having the
optimization at all by 4 or 5% when scanning petstore.cadl thousands of
times, but a little slower than the (broken) bitwise version. This isn't
likely to be noticeable in a full compile, but could be useful if we ever
need to quickly scan for something and skip most of the tokens.