Allow CLI array options to come before positional arguments (#225)
You could not do `cadl compile --import X main.cadl` or `cadl --emit X
main.cadl` because this would be interpreted as importing X and main.cadl or
emitting using X and main.cadl.

The fix is to disable this "greedy array" behavior in yargs. It means that
you have to repeat --import or --emit to specify multiple options, but that
seems fine.