Redirect console.* output in language server to stderr
Logging to stdout breaks LSP when the stdio pipe is used. The failure mode
when console.log is used unexpectedly is really bad and can take a very long
time to diagnose.

This change makes the language server use stderr for all console methods, so
now when decorators or emitters reach for console.log, they will just see
their output.

This change also restores inpsectType and inspectTypeName as decorators. I
think they never got the $ prefix, and I think this was unintentional. They
happen to serve as a good ad-hoc manual test that console.log now works as
we expect, which is why they are reinstated here.

In VS Code, the output to stderr from language server will show up in the
Cadl output pane.

However, in VS Classic, it currently goes to debugger log only. You have to
be attached to see it. It's possible to fix this, but we deleted a whole
bunch of code from the VS extension to stop making output pane messages
ourselves and I'd prefer not to reinstate it unless we decide we really need
it.