Add RouteOptions and AutoRouteOptions to customize generated routes
This change adds a new `RouteOptions` parameter to `getAllRoutes` and
`getRoutesForContainer` which (for now) exposes an `AutoRouteOptions` interface
for customizing how route parameters are generated for `autoRoute` operations.

This enables emitters to customize route generation by allowing route
parameters (the `{param}` syntax) to be renamed or replaced with static
values based on library-specific metadata that is attached to the parameter or
operation.

The caller can also decide whether certain route parameters should
appear in the final set of operation parameters.