Encode OpenAPI refs and remove concept of "ref-safe" names
* Correctly URI-encode OpenAPI refs when necessary. Delete broken code that
attempted to address this with name mangling.

* Remove all regexes from how we decide what to inline and what to ref, and
how to name things. The reasoning is now done on the type objects, and not
on their serialized names.

* The current choices are preserved modulo obvious bugs, but I will open an
issue to revisit some of them. It will be easier to change after this
refactoring.

* Centralize the logic in the shared openapi lib for reuse by cadl-autorest.

* Fix issues with the regex-based stripping of Cadl and service namespaces
by replacing that with a namespace filter callback on Checker.getTypeName.

* Defend against parameter key and type name collisions with diagnostics.