Remove duplicate server state for tracking file types
The separate state could get out of sync with the host state, for example,
our test host server host wasn't sending documentOpen, which caused this.

Rather than fix that and retain future risk of getting out-of-sync a
different way, eliminate the additional state by changing getSourceFileKind
to look for open documents via the same existing host mechanism:
`host.getDocumentByURL`, which is also renamed here to
`getOpenDocumentByURL` for clarity about its purpose. Also, move the
implementation of getSourceFileKind in serverlib closer to the other host
method overrides that have special handling of open documents so that the
pattern is clear.