Fix issue in VS Code on Windows where common defs are not found (#753)
The issue impacts multi-file specs where common definitions are pulled in
from main.cadl but then not seen when editing another file that relies on
that. We started calling realpath in the compiler on relative path imports,
which led to a mismatch between VS Code's lower-case normalized drive letter
and realpath's canonical upper-case drive letter.

This made the language server think that main.cadl did not import the
current file and decide that it should analyze the current file as its own
entry point.

The fix is to also use realpath in the language server whenever we receive a
path or file URL from the client.