Add option to parse JSDoc-like developer doc comments (#1313)
This is currently opt-in via ParseOptions.docs = true. It will need to
become always-on or opt-out when we implement the doc unification proposal
linked to #927, but pending more tests and perf measurement, it seemed
prudent to keep it behind a flag for now. This should unblock website doc
generation consumption while allowing further work in this area to be done
in parallel.

Currently only `@param`, `@template`, and `@returns` (+ alternate spelling
`@return`) are supported. Other tags are parsed as an "unknown" tag
nodes. There is no diagnostic for unknown tags at this time.

`/**...*/` is still allowed in any position as trivia, but can be parsed to
structured data when applied to one of the following nodes:

- any statement
- model property
- operation parameter
- interface operation
- union variant
- enum member

There is also a breaking change here for `#directives` for
consistency. Directives are now also limited to the same set of nodes
above. It seemed rather arbitrary that you could apply them to any list
element and I have never seen them used in other positions like on a
decorator/template/tuple argument.

Single-line `code spans` are recognized as are multi-line code blocks fenced
by triple backticks. Inside these code regions, `@` can be used as content
and will not be parsed as a tag.

Finally, this includes a fix for merge conflict marker scanning. Although we
returned conflict marker tokens, we did not skip the right parts of the
input and therefore did not achieve the desired good error recovery in the
presence of conflict markers.