(Core) Speed up PR validation on windows
We're running up against build timeout on Windows and it's getting in the way
of team productivity.

This change does the following on Windows PR validation:

- Disable docusaurus build
- Disable formatting
- Disable linting

All of these will still be checked on Linux, where the build is currently 2X
as fast, and none of them are likely to find a real Windows-specific issue
in the product.

This change also disables the redundant regen-ref-docs step that is already
performed during website build.