Stop treating models spread into parameters as unreferenced
This was causing confusing duplication in the OpenAPI emit when a model was
used only to spread into operation parameters and the default option to omit
unreferenced types was enabled.

It further caused confusing diffs in the change in progress to use read
visibility as the canonical visibility, so sending this out separately first
to make that easier review.