Remove unnecessary uses of rekeyable maps in projections
Namespace member ordering is implementation defined so this was unnecessary
and rekeyable maps are slower and harder to debug.

The TS typing will force you to use rekeyable map where required. This was
left over from mid-work on member ordering where I had put rekeyable in too
many places in the typing.