Fix development issues after compiler directory move (#2077)
1. Remove unnecessary exclude in tsconfig.json now that we use src/ include
in compiler. Somehow this seemed to fix a rebuild failure for me and it's
consistent with other projects.

2. Fix tmlanguage.js path in watch-tmlanguage

3. Fix compiler path for F5 on VS Code extension

4. Use consistent compiler path for F5 on VS Classic extension. Note that we
should entrypoints/ for F5 since we don't want to run resolution.

5. Delete unused and broken regen-samples in json-schema. It had wrong path
to compiler but even before that fails with non-existent samples dir
already. Also remove compile-samples npm script that just ran a broken tsp
compile command.