Fix JS imports not working when debugging language server
The issue results from cylces in runtime imports, but for reasons that are
not understood yet, this only impacts the language server is run in
development mode and uses entrypoints/server.js directly.

The fix here is the smallest possible: revert a change to an import that was
made inadvertently when compiler source was recently moved under src/.

Will file follow-up issue to lint and/or test to prevent this from happening
again.