Emitter framework async source file (#2183)
Extracted this part of the breaking change from
https://github.com/microsoft/typespec/pull/2157

With prettier 3.0.0 the formatter is async. Which requires the emitter
framework to provide a way to deal with async content mutation.

This only breaks in the emitter framework is calling
`this.emitter.emitSourceFile` directly which would happen if you wrote a
custom `writeOutput`. If usage only include `sourceFile` override then
it still allows returning a non promise.

```diff
-const emittedSf = this.emitter.emitSourceFile(sf);
+const emittedSf = await this.emitter.emitSourceFile(sf);
```