Allow library dependency versions to be specified in init templates (#2212)
This change fixes #1934 by expanding the init template schema to allow
library dependency versions to be specified. I also added a basic unit
test to validate that the `package.json` file gets generated with the
desired versions. More work will be needed to make the scaffolding logic
more testable, but this is a good starting point for now.

For example:

```json
  "with-library-versions": {
    "title": "Test template with library versions",
    "description": "This template specifies two libraries with versions and one without",
    "libraries": [
      {
        "name": "library1",
        "version": "1.0.0"
      },
      {
        "name": "library2",
        "version": "^2.0.0"
      },
      "library3"
    ],
    "skipCompilerPackage": true,
    "config": {},
    "files": [
      {
        "path": "./main.tsp",
        "destination": "main.tsp"
      }
    ]
  }
```

---------

Co-authored-by: Timothee Guerin <timothee.guerin@outlook.com>