Emit diagnostic for an unresolved metatype property reference (#2210)
This PR fixes #2163 by adding a diagnostic which is emitted when a
metatype property reference cannot be resolved. It appears that a
diagnostic was intended to be added here but it was left out!

**Another aspect to this issue:** it appears that metatype properties
are not bound on types until after all members are bound, meaning that
they are not available to be used directly in TypeSpec code. For
example:

```
model A {
  name: string;
}

model B {
  a: A;
}

@test("target")
model Spread {
  ... B.a::type;
}
```

After compiling this file, the `Spread` model will have no properties
because the `type` metatype property of `B.a` cannot be resolved at that
time.

Is that the desired/expected behavior?

---------

Co-authored-by: Timothee Guerin <tiguerin@microsoft.com>