`tsp format` should return non-zero exit code on failures (#2451)
This PR fixes #2323 by returning a non-zero exit code upon any failure
when `tsp format` is executed at the command line. I'm also changing the
return type of `formatTypeSpecFiles` to return a list of file paths that
failed, not sure whether that explicitly qualifies as a breaking change.
If needed, I can create a new function to expose and wrap it with the
original function without changing its signature.