Don't emit `deprecated` for type reference in deprecated declaration (#2478)
This PR fixes #2368 by checking the location where a type reference of a
deprecate type is used and skipping the emit of `deprecated` for that
reference if the parent declaration is also deprecated. This should make
diagnostics for files with multiple deprecations less noisy.

Let me know if I'm missing any other cases!