Website perf improvements (#2507)
Progress in #2480

## `Swc-loader`
Some minor perf improviement. Using `swc-loader` is much faster at
loading the js in the compile stage. This make the timed build nearly
50% faster but it **still gets stuck at** `cache (99%) shutdown
IdleFileCachePlugin` for the vast majority of the time. So this is a
little win but nearly negligable

## Updateing `typedoc-markdown` 

This has changed the file structure to group by folder and has basically
brought back the perf from before the migration to the new doc format.
Not sure why but it works great now.