Emitter framework handle circular references (#2550)
Emitter framework deals with circular reference for you and this works
great if the circular flow involve declarations that can be referenced.
But sometimes there is issues where we can't have a declaration and so a
circular reference becomes impossible. Adding a new `circularReference`
hook that gets called when we resolve a reference that is point back to
itself in the resolution stack. By default it will throw resolve the
declaration ref if the target is a declaration or throw an error.

An emitter can decide to override this function to provide a different
handling in either cases. A target language might not even support
circular reference at all.

---------

Co-authored-by: Mark Cowlishaw <markcowl@microsoft.com>