Fix: Augment decorators might not run if alias caused the type to be resolved too early (#2603)
Issue is that when resolving symbol in the meta types and member
resolution when we resolve an alias or template instance we check the
type. However this makes it that every referenced typed from there on
has already been resolved by the time the augment decorator gets to do
the binding.

A fix for now is to have an option in the symnbol resolution to
checkTemplateTypes or not. We do not check them during the binding phase
but we do allow checking during the checking phase
[#2600](https://github.com/microsoft/typespec/issues/2600)