Emitter framework: Allow reference context to be passed to `emitTypeReference` (#2712)
Current way of setting reference context is flawed in that you cannot
have a different context to different reference. It is also run before
the element itself instead of before emitting the reference so this mean
the context would apply to the referrer as well.

This allows the `emitTypeReference` to take a parameter to patch the
reference context

Next step would be to reconsider the `*ReferenceContext` and `*Context`
methods to make more sense.