VSCode extension: Use rollup for dev build too (#2730)
Problem is vscode still doesn't allow loading es modules so we had all
those hacks to make it work but loading es modules from comon js is
always problematic. Using rollup to generate a commonjs output solve the
problem and we are using that for the published version just not in dev.
Issue with this new approach is `rush watch` won't build the extension
but that's also not something you often need to touch.