Move serialization constructors and raw field (#3414)
This PR migrates the construction of the existing serialization
constructors for a model to the MrwSerializationTypeProvider class. With
this change, all serialization related code is generated with a models
serialization class file. It also adds the support for generating an
additional constructor to deserialize into a model with an additional
raw data parameter.