[protobuf] Handle cross-package references correctly in some buggy cases. (#3561)
Closes #3534
Closes #3556 

This PR adds cross-package resolution in some cases where it was
missing.

In the case of operation inputs, missing effective model type
calculation was causing some input types not to be resolved to external
packages.

In the case of map value types, the calculation of the right hand side
type did not correctly account for the possibility of the type being in
another package, and it now uses the same machinery as other type
references to instrument imports and cross-package references.

The existing scenario test has been extended to prevent regressions and
also extended to test arrays for good measure, though those were working
as expected.

---------

Co-authored-by: Will Temple <will@wtemple.net>
Co-authored-by: Timothee Guerin <tiguerin@microsoft.com>