Expression statement conversion (#3668)
Fixes https://github.com/microsoft/typespec/issues/3499

The idea here is to not have a Statement when all it does is append `;`
to the corresponding Expression. Snippets should return the Expression
if the invoked method returns a value and should return a statement if
the invoked method returns a void.

I added a `Terminate` method to ValueExpression and TypedSnippet to
easily convert to a statement which simply appends the `;`. This is not
an implicit cast since many Expressions cannot be terminated so I wanted
to make this something we do explicitly.