Refactoring typed snippet APIs (#3720)
This pr tackles 3 refactorings.

## Providers that aren't really providers
A ModelProvider is a generic abstraction that allows a plugin to
represent how a model should be shaped. Should it have auto properties
with `{get; set;}` should it have explicit backing fields to track
whether or not a property was every modified to more intelligently
serialize JsonMergePatch? All of these are different strategies that
different ModelProviders can give.

ArgumentProvider on the other hand is an implementation of TypeProvider
that produces a helper class inside the library. It is more of a
definition than a Provider as we would not expect customers to override
this and give a different style for the Argument class.

Because of this the PR changes all such provider classes to have a
suffix of Definition instead.

## TypedSnippet
It is cumbersome to need to always construct a new BinaryDataSnippet
every time you want to access the helper snippets for an expression that
evaluates to a BinaryData. It is also an odd name of a class to have an
instance off since snippets are helpers which are usually static methods
or extension methods. I am adding a `ScopeApi<T>` which we can hang
extension methods off of to allow us to change code like this

```c#
new BinaryDataSnippet(value).ToArray()
```

to this

```c#
value.As<BinaryData>().ToArray()
```

Now any expression can be converted into a type expression and if there
are any snippet extension methods that exist the will appear in
intellisense.

## Static snippet helper
One more area where we need to call expression constructors is in
InvokeStaticMethodConstructor. I have added a helper `Static<T>` which
will change code like this

```c#
new InvokeStaticMethodExpression(typeof(XmlConvert), nameof(XmlConvert.ToString), [timeSpanNoFormat]))
```

to this

```c#
Static<XmlConvert>().Invoke(nameof(XmlConvert.ToString), [timeSpanNoFormat]))
```