Generate optional value type as nullable (#3757)
1. Fix #3740
2. Add `IsNullable` to `PropertyWireInformation` so that downstream
could know whether we should check `null` when doing serialization.

Confirmed after this change, the core part of test code (like
serializing properties) now is exactly same as that in autorest.csharp.
(except some code structure thing like `WriteCore`)(except that in this
this repo there are several newly-added properties like
`FloatExtensibleEnumWithIntValue` which is not in autorest.csharp)

---------

Co-authored-by: Pan Shao <pashao@microsoft.com>
Co-authored-by: m-nash <64171366+m-nash@users.noreply.github.com>