Refactor `MethodProvider` (#3756)
Fixes https://github.com/microsoft/typespec/issues/3739

Previously, constructors and methods are both represented by
`MethodProvider`, which is quite inconvenience in some cases that we
want to get the `ReturnType` out from a method, we will have to cast the
`Signature` to `MethodSignature`.
A new type `ConstructorProvider` is introduced in this PR to represent
ctors only and now the `Signature` type is `MethodSignature` in
`MethodProvider` which only represents the normal methods.

Considering that ctors and methods are similar, we may add a common base
type to `MethodProvider` and `ConstructorProvider`, such as
`MethodBaseProvider` just like how `MethodSignature` and
`ConstructorSignature` do.
But currently I did not see any use case that requires to add ctors and
methods into the same list, therefore this PR does not add that, and we
could always add that part later.

---------

Co-authored-by: m-nash <64171366+m-nash@users.noreply.github.com>