Conversions between parameters, properties and fields (#3754)
Fixes #3692 

In the practice, we usually need to keep tracking the property-parameter
conversion
For instance, when constructing the constructor of a model, we first
construct all the properties of the current model then building up the
ctors. And during building the ctors, we need to convert the properties
to parameters. And in the body of the ctor, we will need to assign the
parameters to the properties, but the code of building signature and
body are in different places, therefore we may lose track of the
conversion relationship therefore we will have to find them by names.

This PR adds the two-way references among `ParameterProvider`,
`PropertyProvider` and `FieldProvider`.
You can convert a property to a parameter by calling the
`PropertyProvider.Parameter`, and the constructed parameter will hold
the reference of the propety which constructs it.
Similarly you could do the same between fields and parameters.
In this way we could prevent that when building something, we will make
things up.

there is also a two-way reference between property and field, but this
is not a conversion, this is to track the field to be the "backing
field" of the property, and vice versa. We do not have a solid use case
for this yet, therefore this part is only prototyping, and if possible,
I could remove it from this PR and add it back later when we have a case
for it (such as the json merge patch case?).

---------

Co-authored-by: m-nash <64171366+m-nash@users.noreply.github.com>